chore: make psycopg2 optional
