+++ requirements.txt
@@ -1,7 +1,6 @@
 numpy~=1.20
 pandas~=1.2
 paramiko==2.7.2
-psycopg2~=2.8.5
 scipy~=1.5
 tqdm==4.29.1
 requests~=2.25
