+++ setup.py
@@ -6,7 +6,6 @@ install_requires = [
     "paramiko",
     "scipy",
     "tqdm",
-    "psycopg2",
     "requests",
 ]
